+++ src/@types/casperlabsSigner.d.ts
@@ -2,11 +2,13 @@ interface CasperLabsHelper {
   /**
    * Returns connection status from Signer
    */
-  isConnected: () => Promise<boolean | undefined>;
+  isConnected: () => Promise<boolean>;
+
   /**
    * Attempt connection to Signer
    */
   requestConnection: () => void;
+
   /**
    * send base16 encoded message to plugin to sign
    *
@@ -14,8 +16,22 @@ interface CasperLabsHelper {
    * @param publicKeyBase64 the base64 encoded public key used to sign the deploy, if set, we will check whether it is the same as the active key for signing the message, otherwise, we won't check.
    */
   sign: (messageBase16: string, publicKeyBase64?: string) => Promise<string>;
-  // returns base64 encoded public key of user current selected account.
-  getSelectedPublicKeyBase64: () => Promise<string | undefined>;
+
+  /*
+   * Returns base64 encoded public key of user current selected account.
+   */
+  getSelectedPublicKeyBase64: () => Promise<string>;
+
+  /**
+   * Retrieve the active public key.
+   * @returns {string} Hex-encoded public key with algorithm prefix.
+   */
+  getActivePublicKey: () => Promise<string>;
+
+  /*
+   * Forces Signer to disconnect from the currently open site.
+   */
+  disconnectFromSite: () => void;
 }
 
 interface SignerTestingHelper {
@@ -30,7 +46,7 @@ interface SignerTestingHelper {
   /**
    * Check if there is an existing vault
    */
-  hasCreatedVault: () => Promise<boolean | undefined>;
+  hasCreatedVault: () => Promise<boolean>;
   /**
    * Reset existing vault (for testing) prevents complications
    * and unlocks in preparation for creating an account