+++ src/lib/Signer.ts
@@ -8,7 +8,7 @@
 /**
  * Check whether CasperLabs Signer extension is connected
  */
-export const isConnected: () => Promise<boolean | undefined> = async () => {
+export const isConnected: () => Promise<boolean> = async () => {
   return await window.casperlabsHelper!.isConnected();
 };
 
@@ -24,11 +24,17 @@ export const sendConnectionRequest: () => void = () => {
  *
  * @throws Error if haven't connected to CasperLabs Signer browser extension.
  */
-export const getSelectedPublicKeyBase64: () => Promise<
-  string | undefined
-> = () => {
-  throwIfNotConnected();
+export const getSelectedPublicKeyBase64: () => Promise<string> = () => {
   return window.casperlabsHelper!.getSelectedPublicKeyBase64();
+};
+
+/**
+ * Retrieve the active public key.
+ *
+ * @returns {string} Hex-encoded public key with algorithm prefix.
+ */
+export const getActivePublicKey: () => Promise<string> = () => {
+  return window.casperlabsHelper!.getActivePublicKey();
 };
 
 /**
@@ -44,8 +50,14 @@ export const sign: (
   messageBase16: string,
   publicKeyBase64?: string
 ) => Promise<string> = (messageBase16: string, publicKeyBase64?: string) => {
-  throwIfNotConnected();
   return window.casperlabsHelper!.sign(messageBase16, publicKeyBase64);
+};
+
+/*
+ * Forces Signer to disconnect from the currently open site.
+ */
+export const disconnectFromSite: () => void = () => {
+  return window.casperlabsHelper!.disconnectFromSite();
 };
 
 export const forceConnection: () => void = () => {
@@ -56,7 +68,7 @@ export const forceDisconnect: () => void = () => {
   return window.signerTestingHelper!.forceDisconnect();
 };
 
-export const hasCreatedVault: () => Promise<boolean | undefined> = () => {
+export const hasCreatedVault: () => Promise<boolean> = () => {
   return window.signerTestingHelper!.hasCreatedVault();
 };
 
@@ -85,12 +97,4 @@ export const signTestDeploy: (msgId: number) => Promise<void> = (
   msgId: number
 ) => {
   return window.signerTestingHelper!.signTestDeploy(msgId);
-};
-
-const throwIfNotConnected = () => {
-  if (!isConnected()) {
-    throw new Error(
-      'No CasperLabs Signer browser plugin detected or it is not ready'
-    );
-  }
 };